+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.1.4] - 2024-12-27
+
 - Use vernier memory usage and rails hooks by default
 - Add support for N+1 detection with prosopite